+++ src/main/java/com/bookmymovie/serviceImpl/UserServiceImpl.java
@@ -1,5 +1,7 @@
 package com.bookmymovie.serviceImpl;
 
+import com.bookmymovie.exceptions.BadRequestException;
+import com.bookmymovie.exceptions.NotFoundException;
 import com.bookmymovie.repository.UserRepository;
 import com.bookmymovie.dto.UserDTO;
 import com.bookmymovie.entity.User;
@@ -26,28 +28,61 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public UserDTO findById(Long id) {
-        return userMapper.userToUserDTO(userRepository.findById(id).orElseThrow(NoSuchElementException::new));
+        if(id==null){
+            throw new BadRequestException("User id cannot be null.");
+        }
+        return userMapper.userToUserDTO(userRepository.findById(id)
+                .orElseThrow(()->new NotFoundException("User with the id of "+id+" not found.")));
     }
 
     @Override
     public void save(User user) {
+        if(user==null){
+            throw new BadRequestException("User cannot be null.");
+        }
         userRepository.save(user);
     }
 
     @Override
     public void update(User user) {
-        User old =userRepository.findById(user.getId()).orElseThrow();
+        if(user==null){
+            throw new BadRequestException("User cannot be null.");
+        }
+        if(user.getRole()=="admin"){
+            throw new BadRequestException("This user is already an admin.");
+        }
+        User old =userRepository.findById(user.getId())
+                .orElseThrow(()->new NotFoundException("User with the id of "+user.getId()+" not found."));
         user.setPassword(old.getPassword());
+        user.setRole("admin");
         userRepository.save(user);
     }
 
     @Override
     public void deleteById(Long id) {
+        if(id==null){
+            throw new BadRequestException("User id cannot be null.");
+        }
+        userRepository.findById(id)
+                .orElseThrow(()->new NotFoundException("User with the id of "+id+" not found."));
         userRepository.deleteById(id);
     }
 
     @Override
+    public boolean doesUserExist(String email) {
+        if(email==null){
+            throw new BadRequestException("User email cannot be null.");
+        }
+        return userRepository.existsByEmail(email);
+    }
+
+    @Override
     public UserDTO findByEmail(String email) {
+        if(email==null){
+            throw new BadRequestException("User email cannot be null.");
+        }
+        userRepository.findByEmail(email)
+                .orElseThrow(()->new NotFoundException("User with the email of "+email+" not found."));
         return userMapper.userToUserDTO(userRepository.findByEmail(email).orElseThrow());
 
     }
